requirements.txt
Django==3.0.3
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
djangorestframework==3.11.0
django-rest-auth==0.9.5
Pillow==7.0.0
psycopg2==2.8.4
stripe==2.42.0
whitenoise==4.1.4
gunicorn==19.9.0
dj-database-url==0.5.0


